requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
et-xmlfile==1.1.0
faiss-cpu==1.8.0.post1
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
# Remove googletrans or adjust its version
# Consider the correct alternative library for translation if needed
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==0.9.1
httplib2==0.22.0
httpx>=0.23.0  # Loosen the version constraint
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-experimental==0.3.0
langchain-google-genai==2.0.0
langchain-text-splitters==0.3.0
langsmith==0.1.121
lxml==5.3.0
MarkupSafe==2.1.5
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pip==24.2
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.1
pydantic_core==2.23.3
pydantic-settings==2.5.2
PyGetWindow==0.0.9
pyparsing==3.1.4
PyPDF2==3.0.1
PyRect==0.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.4
SQLAlchemy==2.0.34
tabulate==0.9.0
tenacity==8.5.0
tqdm==4.66.5
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.4
XlsxWriter==3.2.0
yarl==1.11.1
